+++ processMmr.js
@@ -1,0 +1,28 @@
+var utility = require('./utility');
+var db = require('./db');
+var logger = utility.logger;
+var getData = utility.getData;
+module.exports = function processMmr(job, cb) {
+    var payload = job.data.payload;
+    getData(job.data.url, function(err, data) {
+        if (err) {
+            logger.info(err);
+            return cb(err, err);
+        }
+        logger.info("mmr response");
+        if (data.soloCompetitiveRank || data.competitiveRank) {
+            db.ratings.insert({
+                match_id: payload.match_id,
+                account_id: payload.account_id,
+                soloCompetitiveRank: data.soloCompetitiveRank,
+                competitiveRank: data.competitiveRank,
+                time: new Date()
+            }, function(err) {
+                cb(err);
+            });
+        }
+        else {
+            cb(null);
+        }
+    });
+}
